Added proper verification of webhooks, working on caching mechanism.
The webhook handler now properly verifies the credentials we receive
from plaid. I've started building out a basic interface for retrieving
and caching the key functions we are using for this. Immediately I will
be using the in memory one, but eventually I might move to the redis
one depending on what ends up being more efficient.