Big job/plaid updates.
Moved schema migrations to the internal package. I want to add
functional schema migrations in the future which might involve more than
just database access, but to keep that stuff from being used I want to
keep all of that in the internal package.

I've added much more to the plaid helper interface. This will serve as
an abstraction layer for making API calls to plaid so that we can track
them with sentry traces.

I've migrated all of the jobs to use the new plaid interface rather than
the direct client for now.

I've added an institutions table as well as a job to maintain the basic
institution data for now. It does not keep track of status yet, but will
let us start to add instutition data to the links as they are created in
the future.