Optimizations to contribution calculations.
`Between` is now used to determine the number of contributions for a
given spending object. It is far more efficient than the loop I had
previously. This solves issue #108

I've also made it so the contributions are only recalculated (in regards
to changes in a spending object being paused) when the spending object
goes from a paused state, to an unpaused state. Pausing a spending
object no longer causes a recalculation.