fix: You can no longer select the current date when creating an expense or goal.
When you would select the current date when creating an expense or goal it would result in an error because the time is midnight of the date selected in the user's timezone. Which would always be in the past for the user. This resolves that issue by making the earliest selectable date one day in the future.