feature: Added API endpoint to retrieve institution details via Link ID.
This properly adds a new endpoint to retrieve institution details for
Plaid links via the API. The details are currently sanitized a tiny bit
in that not all of the details are returned. The shape of these details
may change in the future, but this should be fine for now. The entire
institution object is cached in redis so changing response data will not
require other changes.

This also cleans up some UI stuff and introduces the potential for a
"dark mode".