fix: Fixed race condition in pubsub test.
It looks like the listener was being closed before the goroutine that
was using it was properly finished. Resulting in a potential race
condition. This would not be a problem in actual code as the listener is
always closed in the same routine that is is created and used in. But
I've moved the close to the same routine here as well to help fix the
issue. I've also added an atomic counter to still have an assertion that
the correct things happened _after_ the wg.Wait().

Resolves #272