fix: Allow separate Sentry DSN for UI.
This will allow us to more easily support a separate project for the UI
of monetr. But immediately what this lets us do is use a relay server to
send data to sentry. We are not collecting any sensitive data, only
errors from the UI, so this circumvents a lot of ad and tracking
blockers. While that is shitty, in the final product this will be noted
in the privacy policy. As well as documented publically. Again, no
personal data is being sent to sentry. Only errors and information to
diagnose errors.