fix: Upgrading to react router v6 + more.
Upgrading everything to react-router v6. But will probably switch to
wouter in the future given how tedious react-router has become.

This also makes several other changes. It furthers our migration to
functional components and using react hooks. It also completely removes
react-hot-loader.

On the backend the authentication approach has been changed as well. The
`M-Token` cookie is now `HttpOnly` which does not allow the UI to see it
at all. This is a huge security improvement; but requires some changes
in how authentication is performed by the frontend.

Now the frontend will make a request to `/users/me` upon bootstrapping.
If the request succeeds, then the frontend assumes the token is present
(even if invisible to the UI code) and proceeds to bootstrap
authentication. If the request fails, the UI assumes that the user is
not authenticated and falls back to the login page.

On the backend, the cookie is read for each authenticated request. If
the cookie is not valid for a request. The response for that request
will remove the cookie.

Because the UI cannot directly interact with the cookie, an
`/authentication/logout` path has been added to force removal of the
authentication token.