feat: Laying ground-work for manual syncing.
In the future I want user's to be able to periodically trigger a Plaid sync manually if something isn't working quite right. This adds a column that will be used to determine the last time they manually triggered a sync. This way we can limit how often a user can manually sync.