fix: Prevent multiple subscriptions from being made.
When a subscription is "past_due", right now it is possible for another
subscription to be made. With the updated payment method; both
subscriptions can then charge the customer. This is an attempt to
resolve that by preventing a new subscription from being made unless the
original subscription is truely canceled or expired.

Resolves #717