feat(spending): Support more frequent spending than funding.
This adds support for spending objects that recur more frequently than
they are funded. It will over allocate a spending object upon each
funding schedule, thus giving it enough to satisfy each recurrence of
the spending.

This also adds a cron job to clean up stale spending objects that don't
get spent from. If a spending object passes its due date and is not
spent, it should be bumped to make sure that it has enough ready for the
next due date. This will not typically effect normal expenses, but will
play a role in making sure that more frequent expenses are calculated
properly.

Resolves #141