feat(funding): Allow funding schedules to exclude weekends.
Right now if a funding schedule falls on a weekend it is likely that the
pay will actually be deposited into the person's account the prior
business day. Because of this; funding schedules can process late for
pay cadences that do not always fall on week days.

This adds a setting to the funding schedule object to allow weekends to
be excluded from the contribution dates. The funding schedule itself
will still incremenent normally, but if the contribution date were to
fall on a weekend, then the date will be adjusted to the closest
previous business day.