feat(cli): Adding tool to export your data to JSON.
Working on a way for people to be able to export all of their monetr
data to a JSON file. This file can then be used for their own purposes,
or will soon be able to be imported back into another monetr instance.
Allowing people to use their real budgeting data for things like local
development easily.