fix(ui): Improved `Cache-Control` of UI content files.
Right now we are only serving UI files with an expires header, this is
not enough for us to achieve a good caching of the UI especially when
changes are made. This adds a cache control header with a TTL of 24
hours to make sure that UI content is refreshed at a reasonable
frequency.