feat(forecast): Laying groundwork for forecasting funding.
This provides a new interface that can be used to calculate funding
events. Funding events are when a funding schedule should be triggered.
This interface allows easy wrapping of multiple funding schedules to
interlace funding events.

This code should not be relied on yet, it is still a work in progress.