feat(api): Adding basic support for mobile authentication.
Mobile clients will not be using an HTTP only cookie, instead they will
need to derive the token from the response directly and store it.

This adds `isMobile` as a valid field for the login request, and adds
parsing of the token header (with the same name as the cookie) to the
authentication middleware.