feat(ui): Next funding contribution is based on forecasting.
The next contribution amount shown in the funding schedules view is now
based on the forecasting endpoint. This means that the contribution is
based on both the amount needed for expenses now, as well as the amount
spent from expenses between now and the next funding event. This should
provide a (generally) more accurate prediction on next contribution.