fix(ui): Fixed Plaid dialog closing when updating/reauthenticating.
Because of the way the dialogs were being rendered, this component could
get unmounted because I suck at this game.

This resolves the problem by moving the dialog and its rendering into
NiceModal which is mounted much higher in the component tree, preventing
it from being trampled by odd re-renders.

For real this time:
Resolves #1205