fix(forecasting): Fix forecasting timeouts (#1317)
* fix(forecast): Fixing forecasting timing out.

The forecasting code would get stuck in a loop when goals met a certain
criteria. This adds tests to assert that the same thing wont happen
again in the future as well as fixes the problem.

Resolves #1243

* chore: Adding safety check for detecting this bug in the future.

If this bug does come up again this will give some insight into how to
fix it.