fix(ui): Fixed hot module replacement woes for css and rspack.
When I was making updates that would require tailwind to re-render the
css file; it would not consistently persist via the HMR. But reloading
the page caused it to work just fine. Turns out this was just a
configuration error anyway. But this resolves it and bumps some
dependencies and adds some niceties for running things.