fix(forecast): Fixed forecast timing bug.
Fixed an issue where funding events could be missed depending on the
timezone and timestamp provided for the forecasting date range. This was
caused by an incorrect conversion of the input timestamp to a midnight
timestamp. Causing that midnight timestamp to actually be in the future
relative to the input. This issue has been resolved and the faulty
midnight function has been removed. A new midnight function was added
that correctly builds the timestamps.

Resolves #1486